First version that inserts into Cassandra.
