--- server_java/.gitignore
@@ -1,2 +1,0 @@
-classes
-files/dummy-web-server.jar
