+++ server_java/files/start.sh
@@ -1,5 +1,3 @@
 #!/bin/sh
 
-THEDB=${THEDB:-reto1db}
-
-java -cp /root/mysql-connector-java-5.0.8-bin.jar:/root/dummy-web-server.jar com.happylife.demo.Main $THEDB
+java -cp /root/lib/*:/root/web-server.jar net.bluetab.bigdataactivity.http_server.Main
