--- server_java/src/com/happylife/demo/Main.java
@@ -1,9 +1,0 @@
-package com.happylife.demo;
-
-public class Main {
-   public static void main(String[] args) {
-      // start http server
-      SimpleHttpServer httpServer = new SimpleHttpServer();
-      httpServer.Start();
-   }
-}
