+++ server_java/src/main/java/net/bluetab/bigdataactivity/http_server/Main.java
@@ -1,0 +1,9 @@
+package net.bluetab.bigdataactivity.http_server;
+
+public class Main {
+   public static void main(String[] args) {
+      // start http server
+      CassandraLogHttpServer httpServer = new CassandraLogHttpServer();
+      httpServer.start();
+   }
+}
